Test/feature/upgrade (#89)
* Compilation and dependency issues fix #1

* Vurnabilities Fix #1

* Vurnabilities fix #2

---------

+++ src/main/java/com/iemr/hwc/repo/doctor/PrescriptionTemplatesRepo.java
@@ -32,7 +32,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Repository
-public interface PrescriptionTemplatesRepo extends CrudRepository<PrescriptionTemplates,Integer> {
+public interface PrescriptionTemplatesRepo extends CrudRepository<PrescriptionTemplates,Long> {
 
     @Query(" SELECT t FROM PrescriptionTemplates t WHERE t.userID = :userId")
     List<PrescriptionTemplates> getPrescriptionTemplatesByUserID(@Param("userId") Integer userId);
