+++ src/main/java/com/iemr/hwc/repo/masterrepo/nurse/CancerPersonalHabitMasterRepo.java
@@ -32,7 +32,7 @@ import org.springframework.stereotype.Repository;
 import com.iemr.hwc.data.masterdata.nurse.CancerPersonalHabitType;
 @Repository
 @RestResource(exported = false)
-public interface CancerPersonalHabitMasterRepo extends CrudRepository<CancerPersonalHabitType, Long>{
+public interface CancerPersonalHabitMasterRepo extends CrudRepository<CancerPersonalHabitType, Short>{
 
 	@Query("select cancerPersonalHabitID, habitType, habitValue from CancerPersonalHabitType where deleted = false and habitType=:habitType "
 			+ "order by habitValue ")
